main.jsをわけた
